feat(navigation): refactor navigation components and replace CourseNavigation with Navigation for improved structure
